functions/kubeless/image-to-text/requirements.txt
minio
pillow
pytesseract
jaeger-client
opentracing
requests
bottle